Merged PR 5389: 24066-Fix-Name-of-employer-field-is-not-disabled-on-edit-modal
Verify the 'Name of Employer' field.

Expected Result
It is a text field and mandatory.
The system displayed a validation message as "Name of Employer is required".
Comments: 1. Name of Employer field is not disabled so the user can edit this field and add any employer.

Attachment(s):
Screenshot-2024-04-23T06

Related work items: #24066