Merged PR 5396: 24208---Name-fields-should-allow-special-characters
NOTE:  THIS BUG WAS FIXED IN A PREVIOUS RELEASE FOR ADD BUT NOT EDIT.
The name fields in the Payment Address modal need to allow for special characters just like the bug fix referenced here. See attachments for examples.

Some of the special characters listed in the referenced bug are:
 /  ( currently 242 instances in the Address Line 2) Typically used for  "c/o "
#  (currently 207 instances in the Address Line 2)
:  (currently 19instance in the Address Line 2)
" (currently 1 instance in the Address Line 2)
() (currently 7 instances in the Address Line 2)
- (dash) (currently 60instances in the Address Line 2)
& (currently 3 instances in the Address Line 2)
' (currently no information regarding the number of times this is present in the data)

Related work items: #24208